Add namespace and use class scope for VObjectType enum
VObjectType is now a public header and it was missing the namespace.
Also, the lack of use of enum class generated tons of static analyzer
warnings in client code.

Fixed both issues - use enum class and wrapped it in namespace.

+++ src/Cache/ParseCache.cpp
@@ -250,7 +250,7 @@ bool ParseCache::save() {
           elem->m_type, (RawNodeId)elem->m_uniqueId, elem->m_line,
           elem->m_column, elem->m_endLine, elem->m_endColumn, timeInfo,
           (RawNodeId)elem->m_parent, (RawNodeId)elem->m_node,
-          elem->m_defaultNetType));
+          static_cast<unsigned int>(elem->m_defaultNetType)));
     }
   auto elementList = builder.CreateVector(element_vec);
